refactor settings and add report
