+++ src/app.router.js
@@ -1,7 +1,9 @@
 import { useRoutes } from 'react-router-dom';
 import Home from './pages/home';
+import TimerSettings from './pages/settings';
 import Login from './pages/login';
 import PomodoroPage from './pages/pomodoro';
+import Report from './pages/report';
 import SignUp from './pages/signup';
 import Tasks from './pages/tasks';
 
@@ -17,6 +19,14 @@ const publicRoutes = [
       {
         path: 'tasks',
         element: <Tasks />,
+      },
+      {
+        path: 'settings',
+        element: <TimerSettings />,
+      },
+      {
+        path: 'report',
+        element: <Report />,
       },
     ],
   },
