+++ src/pages/home/HomePage.jsx
@@ -1,4 +1,4 @@
-import { useDisclosure } from '@chakra-ui/react';
+import { Box, useDisclosure } from '@chakra-ui/react';
 import React from 'react';
 import { Outlet } from 'react-router-dom';
 import { useTimer } from 'src/contexts';
@@ -6,12 +6,10 @@ import { useAuth } from 'src/contexts';
 import { AuthService } from 'src/services';
 import Footer from './components/footer';
 import Header from './components/header';
-import TimerSettings from './components/settings';
 
 const HomePage = () => {
-  const { isOpen, onOpen, onClose } = useDisclosure();
-  const { updateSettings, time } = useTimer();
   const { authState, authDispatch } = useAuth();
+
   const handleLoginWithGoogle = async () => {
     const result = await AuthService.authWithGoogle().catch(e => {
       console.log(e.toString());
@@ -27,21 +25,17 @@ const HomePage = () => {
   };
 
   return (
-    <div>
+    <Box>
       <Header
         onLogin={handleLoginWithGoogle}
         onLogout={handleLogout}
-        onSettings={onOpen}
         isLoggedIn={authState.isLoggedIn}
       />
-      <Outlet />
-      <TimerSettings
-        isOpen={isOpen}
-        onClose={onClose}
-        updateSettings={updateSettings}
-      />
+      <Box minHeight={'calc(100vh - 140px)'}>
+        <Outlet />
+      </Box>
       <Footer />
-    </div>
+    </Box>
   );
 };
