+++ src/pages/report/Report.jsx
@@ -1,0 +1,8 @@
+import { Heading } from '@chakra-ui/react';
+import React from 'react';
+
+function Report() {
+  return <Heading m={'10rem'}> Coming soon !</Heading>;
+}
+
+export default Report;
