+++ src/pages/report/index.js
@@ -1,0 +1,3 @@
+import Report from './Report';
+
+export default Report;
