+++ src/pages/settings/index.js
@@ -1,0 +1,3 @@
+import TimerSettings from './settings';
+
+export default TimerSettings;
